# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: KialaBet Sportsbook API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /wallet/deposit:
    post:
      operationId: depositFunds
      tags: [wallet]
      summary: Demo deposit (no real payment)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepositInput"
      responses:
        "200":
          description: Deposit simulated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Wallet"
  /wallet/transactions:
    get:
      operationId: listTransactions
      tags: [wallet]
      summary: Get transaction history
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"

  # JACKPOTS
...
  /profile:
    get:
      operationId: getProfile
      tags: [users]
      summary: Get user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateProfile
      tags: [users]
      summary: Update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /notifications/{notificationId}/read:
    post:
      operationId: markNotificationRead
      tags: [users]
      summary: Mark notification as read
      parameters:
        - name: notificationId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Marked read
...
components:
  schemas:
...
    User:
      type: object
      required: [id, username, email, phone, role, isSuspended, createdAt]
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [user, admin]
        isSuspended:
          type: boolean
        referralCode:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        createdAt:
          type: string
    ProfileUpdate:
      type: object
      properties:
        username:
          type: string
        phone:
          type: string
        avatarUrl:
          type: ["string", "null"]

    # MATCH SCHEMAS
...
    Wallet:
      type: object
      required: [id, userId, balance, bonusBalance, currency]
      properties:
        id:
          type: integer
        userId:
          type: integer
        balance:
          type: number
        bonusBalance:
          type: number
        currency:
          type: string
    DepositInput:
      type: object
      required: [amount, method]
      properties:
        amount:
          type: number
        method:
          type: string
          enum: [mpesa, airtel, card, demo]
    Transaction:
      type: object
      required: [id, type, amount, status, description, createdAt]
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [deposit, withdrawal, bet, win, bonus, refund]
        amount:
          type: number
        status:
          type: string
          enum: [pending, completed, failed]
        description:
          type: string
        createdAt:
          type: string

    # JACKPOT SCHEMAS
...